requirements.txt
requests>=2.0
python-telegram-bot>=20.0
